feat：尝试拆解monitoring
1.规范组件文件分类
2.将login转为script setup写法
3.尝试拆解monitoring
