+++ src/components/common/TitleBar/logTitleBar.vue
