+++ src/components/monitoring/Dialog/controlDialog.vue
