+++ src/components/monitoring/Dialog/deleteDialog.vue
