+++ src/components/monitoring/Dialog/intelligentControlDialog.vue
