+++ src/pages/login.vue
@@ -32,7 +32,7 @@
   </div>
 </template>
 
-<script>
+<script setup>
 import { ref, onMounted } from "vue";
 import { useCustomStore } from '@/store'; // 引入pinia
 import { post } from "../api/http.js";
@@ -40,118 +40,107 @@ import { ElMessage } from "element-plus";
 import { JSEncrypt } from "jsencrypt";
 import { useIpcRenderer } from "@vueuse/electron";
 
-export default {
-  name: "Login",
-  setup() {
-    const username = ref("");
-    const password = ref("");
-    let recordPassword = ref(false);
-    const electronStore = require('electron-store');
-    const Estore = new electronStore();
-    const store = useCustomStore();
-    const ipcRenderer = useIpcRenderer();
+const username = ref("");
+const password = ref("");
+let recordPassword = ref(false);
+const electronStore = require('electron-store');
+const Estore = new electronStore();
+const store = useCustomStore();
+const ipcRenderer = useIpcRenderer();
 
-    // 加密函数
-    function encryptPWD(password) {
-      const encryptor = new JSEncrypt()
-      const key = "MFwwDQYJKoZIhvcNAQEBBQADSwAwSAJBALiY4DWLKMiLRypsz4A8zdWyQF1fjZGof66PkXDt1AAwuEoKo8rNG9oP5iMZYr0TRDflCUebkP384qzDUFqcA3cCAwEAAQ=="
-      encryptor.setPublicKey(key)
-      return encryptor.encrypt(password + '')
-    }
+// 加密函数
+function encryptPWD(password) {
+  const encryptor = new JSEncrypt()
+  const key = "MFwwDQYJKoZIhvcNAQEBBQADSwAwSAJBALiY4DWLKMiLRypsz4A8zdWyQF1fjZGof66PkXDt1AAwuEoKo8rNG9oP5iMZYr0TRDflCUebkP384qzDUFqcA3cCAwEAAQ=="
+  encryptor.setPublicKey(key)
+  return encryptor.encrypt(password + '')
+}
 
-    // 尝试登录
-    function tryLogin() {
-      const customUrl = "http://lab.zhongyaohui.club/login";
-      const postData = {
-        username: username.value.trim(),
-        password: encryptPWD(password.value.trim()),
-      };
-      if (postData.username === "" || postData.password === "") {
-        ElMessage({
-          showClose: true,
-          message: "输入的内容不能为空！",
-          type: "error",
-          offset: 360
-        });
-      } else {
-        post(customUrl, postData)
-          .then(handleLoginResponse)
-          .catch((error) => {
-            console.error("登录失败:", error);
-          });
-      }
-    }
-
-    // 处理登录响应
-    function handleLoginResponse(res) {
-      if (res.code === 21200) {
-        const userData = res.data;
-
-        // 如果“记住密码”被选中，保存凭据到本地存储
-        if (recordPassword.value) {
-          Estore.set("logindata", {
-            username: username.value,
-            password: password.value,
-          });
-          Estore.set("recordPassword", recordPassword.value);
-        } else {
-          // 如果“记住密码”未选中，清除本地存储的凭据
-          Estore.delete("logindata");
-          Estore.set("recordPassword", recordPassword.value);
-        }
-
-        store.setUserdata(userData)
-
-        // 只在第一次登录/token过期后/切换账号时 存入token
-        if (!Estore.get('token')) {
-          console.log('只在第一次登录/token过期后/切换账号时 存入token');
-          // 登录成功后将 token 存入本地存储
-          const token = userData.token;
-          Estore.set("token",token);
-        }
-
-        // 跳转到 '/monitoring' 页面
-        ipcRenderer.send("login-access");
-      } else {
-        ElMessage({
-          showClose: true,
-          message: "用户名或密码错误！",
-          type: "error",
-          offset: 360
-        });
-      }
-    }
-
-    // 跳转至找回密码
-    function findPWD(){
-      ipcRenderer.send("findPWD-open")
-    }
-
-    // 记住密码  在组件挂载时从本地存储加载保存的凭据并自动登录
-    onMounted(() => {
-      recordPassword.value = Estore.get('recordPassword')
-      // 判断是否记住密码
-      if (Estore.get('recordPassword')) {
-        let logindata = Estore.get('logindata')
-        let savedUsername = logindata.username
-        let savedPassword = logindata.password
-        // 确保保存的信息有值
-        if (savedUsername && savedPassword) {
-          username.value = savedUsername;
-          password.value = savedPassword;
-        }
-      }
+// 尝试登录
+function tryLogin() {
+  const customUrl = "http://lab.zhongyaohui.club/login";
+  const postData = {
+    username: username.value.trim(),
+    password: encryptPWD(password.value.trim()),
+  };
+  if (postData.username === "" || postData.password === "") {
+    ElMessage({
+      showClose: true,
+      message: "输入的内容不能为空！",
+      type: "error",
+      offset: 360
     });
+  } else {
+    post(customUrl, postData)
+      .then(handleLoginResponse)
+      .catch((error) => {
+        console.error("登录失败:", error);
+      });
+  }
+}
 
-    return {
-      username,
-      password,
-      recordPassword,
-      tryLogin,
-      findPWD
-    };
-  },
-};
+// 处理登录响应
+function handleLoginResponse(res) {
+  if (res.code === 21200) {
+    const userData = res.data;
+
+    // 如果“记住密码”被选中，保存凭据到本地存储
+    if (recordPassword.value) {
+      Estore.set("logindata", {
+        username: username.value,
+        password: password.value,
+      });
+      Estore.set("recordPassword", recordPassword.value);
+    } else {
+      // 如果“记住密码”未选中，清除本地存储的凭据
+      Estore.delete("logindata");
+      Estore.set("recordPassword", recordPassword.value);
+    }
+
+    store.setUserdata(userData)
+
+    // 只在第一次登录/token过期后/切换账号时 存入token
+    if (!Estore.get('token')) {
+      console.log('只在第一次登录/token过期后/切换账号时 存入token');
+      // 登录成功后将 token 存入本地存储
+      const token = userData.token;
+      Estore.set("token", token);
+    }
+
+    // 跳转到 '/monitoring' 页面
+    ipcRenderer.send("login-access");
+  } else {
+    ElMessage({
+      showClose: true,
+      message: "用户名或密码错误！",
+      type: "error",
+      offset: 360
+    });
+  }
+}
+
+// 跳转至找回密码
+function findPWD() {
+  ipcRenderer.send("findPWD-open")
+}
+
+// 记住密码  在组件挂载时从本地存储加载保存的凭据并自动登录
+onMounted(() => {
+  recordPassword.value = Estore.get('recordPassword')
+  // 判断是否记住密码
+  if (Estore.get('recordPassword')) {
+    let logindata = Estore.get('logindata')
+    let savedUsername = logindata.username
+    let savedPassword = logindata.password
+    // 确保保存的信息有值
+    if (savedUsername && savedPassword) {
+      username.value = savedUsername;
+      password.value = savedPassword;
+    }
+  }
+});
+
+
 </script>
 
 
@@ -160,7 +149,7 @@ export default {
 #loginBox {
   width: 375px;
   height: 480px;
-  border: 1px solid rgb(217,217,218);
+  border: 1px solid rgb(217, 217, 218);
   border-radius: $border-radius;
   background-color: #fff;
   position: absolute;
@@ -198,7 +187,7 @@ export default {
   }
 }
 
-#main{
+#main {
   margin-top: 35px;
 }
